requirements.txt
dnspython
speedtest-cli
requests
ping3
python-whois
cryptography
scapy
colorama
ipwhois
defusedxml